refactor(agent): extract YAML frontmatter parsing from discovery module
