+++ packages/agent/test/agent/discovery.test.ts
@@ -1,6 +1,7 @@
 import { describe, it, expect, beforeEach } from "bun:test";
 import { join } from "node:path";
-import { AgentDiscovery, parseFrontmatter } from "../../src/agent/discovery";
+import { AgentDiscovery } from "../../src/agent/discovery";
+import { parseFrontmatter } from "../../src/agent/frontmatter";
 import { BuiltinAgentRegistry } from "../../src/agent/registry";
 
 const FIXTURES_DIR = join(import.meta.dir, "../fixtures/agents");
